chore: Inline integrations routes into the main router file
